requirements.txt
httplib2==0.22.0
magicmock==0.3
pyparsing==3.2.0
RPi.GPIO==0.7.1
